requirements.txt
langchain==0.1.0
openai==1.7.0
chromadb>=0.4.18
tiktoken>=0.5.2
pydantic>=2.5.2
litellm>=1.10.2
ipykernel>=6.27.1
langchain-openai==0.0.2
boto3>=1.33.2,<1.34.35
aioboto3==12.3.0

python-dotenv~=1.0.1
requests~=2.32.3
uvicorn~=0.30.1
fastapi~=0.111.1